[Update] 날씨를 request로 받아오는게 아니라 현재 시각 날씨로 자동 추천

+++ src/main/java/Nanali/repository/OutfitRepository.java
@@ -17,8 +17,8 @@ public interface OutfitRepository extends JpaRepository<Outfit, Long> {
             "and o.sex = :sex or o.temp_from <= :temp and o.temp_to >= :temp and o.uv_from <= :uv " +
             "and o.uv_to >= :uv and o.rain_from <= :rain and o.rain_to >= : rain and o.style = :style " +
             "and o.sex = 'BOTH'")
-    List<Outfit> findAllOutifs(@Param(value = "temp") Long temp, @Param(value = "uv") Long uv,
-                               @Param(value = "rain") Long rain, @Param(value = "style") Style style, @Param(value = "sex") Sex sex);
+    List<Outfit> findAllOutifs(@Param(value = "temp") Double temp, @Param(value = "uv") Double uv,
+                               @Param(value = "rain") Double rain, @Param(value = "style") Style style, @Param(value = "sex") Sex sex);
     // Style 타입으로 수정 필요
 
     Optional<Outfit> findById(Long id);
